reaction widget update 02-04-25

+++ lib/feature/domain_layer/entity/obsparameter_entity.dart
@@ -51,7 +51,7 @@ class ObservationByParameterEntity extends Equatable {
     final String? iqcIspParamName;
     final int? iqcIioId;
     final String? iqcIioObservationNotes;
-    final int? iqcIioObservationNumericValue;
+    final double? iqcIioObservationNumericValue;
     final DateTime? iqcIioDate;
     final double? iqcCpsRangeFrom;
     final int? iqcIisSampleSno;
